feat(client): :construction: Display list of needs
