+++ client/src/components/Pages/recherche/ThemeMenu/Need.tsx
@@ -1,4 +1,4 @@
-import CheckboxIcon from "components/UI/CheckboxIcon";
+import Checkbox from "components/UI/Checkbox";
 import React from "react";
 import styles from "./Need.module.css";
 
@@ -8,13 +8,12 @@ interface Props {
 
 const Need: React.FC<Props> = ({ label }) => {
   return (
-    <div className={styles.container}>
-      <CheckboxIcon />
+    <Checkbox>
       <span className={styles.label}>{label}</span>
       <div className={styles.countContainer}>
         <div className={styles.count}>{133}</div>
       </div>
-    </div>
+    </Checkbox>
   );
 };
